+++ src/app/feature/comic/shared/services/comic.service.ts
@@ -10,20 +10,11 @@ import { environment } from 'src/environments/environment';
 export class ComicService {
 
   constructor(
-    private http: HttpService,
+    private http: HttpService
   ) { }
 
-  getRandom(dateRange:string,limit = 3): Observable<Http> {
-    let params = {
-      'limit': limit,
-      'dateRange': dateRange
-    };
-    return this.http.doGet(`${environment.api}v1/public/comics`,{params})
+  get(id: string): Observable<Http> {
+    return this.http.doGet(`${environment.api}/v1/public/characters/${id}/comics`)
   }
-
-  getOne(uri: string) {
-    return this.http.doGet(`${uri}`)
-  }
-
 
 }