+++ src/app/feature/home/components/card/card.component.ts
@@ -1,7 +1,10 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
 import { Character } from '@home/shared/model/character.model';
 import { CharacterService } from '@home/shared/services/character/character.service';
+import { ComicService } from '@home/shared/services/comic/comic.service';
+import { FavoriteService } from '@home/shared/services/favorite/favorite.service';
 import { ModalComponent } from '@shared/components/modal/modal.component';
+import { AlertService } from '@shared/services/alert.service';
 import { fromEvent } from 'rxjs';
 
 @Component({
@@ -18,15 +21,10 @@ export class CardComponent implements OnInit {
 
   constructor(
     private apiCharacter: CharacterService,
+    private apiComic: ComicService,
+    private apiFavorite: FavoriteService,
+    private alertService: AlertService,
   ) { }
-
-  // ngOnChanges(changes: SimpleChanges): void {
-  //   console.log(changes);
-  //   this.helper.send.subscribe(data => this.search = data);
-  //   if(!this.search) {
-  //     this.doSearch();
-  //   }
-  // }
 
   ngOnInit(): void {
     this.get();
@@ -46,6 +44,25 @@ export class CardComponent implements OnInit {
       this.record = character;
       this.modal.open(this.record, false, false);
     });
+  }
+
+  getOneComic(uri: string) {
+    this.apiComic.getOne(uri).subscribe(response => {
+      console.log(response);
+      const [comic] = response?.data?.results;
+      this.record = comic;
+      this.modal.open(this.record, true, true);
+    });
+  }
+
+  add(data: any) {
+    console.log(data.id);
+    if(this.apiFavorite.add(data)) {
+      this.alertService.successAlert('The comic was added successfuly');
+      this.modal.close();
+    } else {
+      this.alertService.infoAlert('The comic can\'t added beacause it\'s already in the list')
+    }
   }
 
   doSearch() {